Generalize download csv code
