+++ packages/client/src/components/common/buttons/index.ts
@@ -6,6 +6,7 @@ export * from './confirm-mini-button.js';
 export * from './copy-to-clipboard-button.js';
 export * from './delete-charge-button.js';
 export * from './delete-document-button.js';
+export * from './download-csv-button.js';
 export * from './edit-mini-button.js';
 export * from './info-mini-button.js';
 export * from './insert-mini-button.js';
